refactor: Extract runtime parsing and schema to "runtime" file
