+++ src/domain/query-states.spec.ts
@@ -5,7 +5,7 @@ import { act } from "react";
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
 import { z } from "zod";
 
-import { runtimeEnv } from "src/env/client";
+import { runtimeEnv } from "src/env/runtime";
 
 import { makeLinkGenerator } from "../../src/lib/use-query-state";
